third_party/blink/web_tests/pending_beacon/pending_beacon-deactivate.html

<!DOCTYPE html>
<head>
<meta charset="utf-8">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<script>
  const BeaconTypes = [
    {type: PendingPostBeacon, name: "PendingPostBeacon", expectedMethod: "POST"},
    {type: PendingGetBeacon, name: "PendingGetBeacon", expectedMethod: "GET"},
  ];

  for (const beaconType of BeaconTypes) {
    test(() => {
      const beacon = new beaconType.type('https://www.google.com');
      assert_true(beacon.pending);
      beacon.deactivate();
      assert_false(beacon.pending);
    }, `${beaconType.name}: Tests deactivate()`);
  }
</script>
</body>
</html>
